Organize store folder
