+++ src/app/gallery/Breadcrumps.spec.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import { Provider } from "react-redux";
 import { createItemsBasedOnStructure } from "../scenarious/helpers/itemsBuilder";
-import { store, actions } from "../state/store";
+import { store, actions } from "../state";
 import Breadcrumps from "./Breadcrumps";
 
 describe("Having a sidebar with some folders", () => {
